\name{emdbook-package}
\alias{emdbook-package}
\alias{emdbook}
\docType{package}
\title{
ecological models and data
}
\description{
Data sets and auxiliary functions for \emph{Ecological Models and Data}
}
\details{
\tabular{ll}{
Package: \tab emdbook\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-05-30\cr
License: \tab GPL\cr
}
Data sets (accessible via data()) and auxiliary functions
}
\author{
Ben Bolker

Maintainer: Ben Bolker <bolker@zoo.ufl.edu>
}
\references{
Ecological Models and Data
}
\keyword{ package }
