\name{get.emdbook.packages}
\alias{get.emdbook.packages}
\alias{update.bmb.packages}
\title{install and update auxiliary packages}
\description{
   convenience function for downloading and installing
   all the packages needed for the book (just a list
   and a wrapper around \code{\link{install.packages}})
}
\usage{
get.emdbook.packages()
}
\value{
  none: installs packages as a side effect
}
\author{Ben Bolker}
\seealso{\code{\link{install.packages}}}
\keyword{misc}
